+++ platform/app/src/server/api/root.ts
@@ -78,6 +78,7 @@ import { subscriptionRouter } from "./routers/subscription";
 import { suiteRouter } from "./routers/suites";
 import { teamRouter } from "./routers/team";
 import { topicsRouter } from "./routers/topics";
+import { traceEditOverlayRouter } from "./routers/traceEditOverlay";
 import { tracesRouter } from "./routers/traces";
 import { tracesV2Router } from "./routers/tracesV2";
 import { translateRouter } from "./routers/translate";
@@ -95,6 +96,7 @@ const coreRouters = {
   team: teamRouter,
   traces: tracesRouter,
   tracesV2: tracesV2Router,
+  traceEditOverlay: traceEditOverlayRouter,
   codingAgents: codingAgentsRouter,
   spans: spansRouter,
   analytics: analyticsRouter,